fix: createFilter runtime
