+++ src/runtime/plugin/view/client.ts
@@ -3,7 +3,7 @@ import { computed, createApp, h, ref, shallowReactive, unref } from 'vue'
 import type { NuxtApp } from 'nuxt/app'
 import type { NuxtDevtoolsIframeClient } from '@nuxt/devtools-kit/types'
 import type { LinkInspectionResult, NuxtLinkCheckerClient } from '../../types'
-import { createFilter } from '../../../urlFilter'
+import { createFilter } from '../../sharedUtils'
 import Main from './Main.vue'
 import { linkDb } from './state'
 import { useRoute, useRuntimeConfig } from '#imports'
